+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version (development version)
+
+ * ...
+ 
+
 # Version 0.16.2 [2022-11-21]
 
 ## Documentation